repackaged cohortfinder to expose the API
