+++ cohortfinder/__init__.py
@@ -1,3 +1,5 @@
+from .cohortfinder import runCohortFinder
+
 try:
     from ._version import version as __version__
 except ImportError:  # pragma: no cover
